fileman-main/requirements.txt
Flask
flask-login
flask_sqlalchemy
flask_bcrypt
psycopg2-binary
python-dotenv
flask_session
python-magic
beautifulsoup4